requirements.txt
flask
sqlalchemy
pandas
numpy
requests
scikit-learn
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==2.4.1
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
comm==0.2.2
curanet==0.1.0
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
deepseek==1.0.0
distro==1.9.0
docutils==0.21.2
executing==2.1.0
fastapi==0.115.8
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
id==1.5.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-deepseek==0.0.1
langchain-deepseek-official==0.1.0
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.3.3
llm==0.20
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nh3==0.2.20
numpy==1.26.4
openai==1.60.2
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-ulid==3.0.0
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
stack-data==0.6.3
starlette==0.45.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.8.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
twine==6.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
zstandard==0.23.0

